Update the crc_utility functions for easier verification
These functions used larger counter types than strictly necessary. These have
been reduced to types of more appropriate sizes. Additionally, it was
discovered that a signed loop counter was declared and compared against an
unsigned end condition. This mismatch was fixed, though it would not have
resulted in incorrect behavior on any remotely sane system as used.

+++ include/internal/crc_utility.h
@@ -1,6 +1,6 @@
 #ifndef SIMPLECRC_UTILITY_H_
 #define SIMPLECRC_UTILITY_H_
 
-uint64_t reflect(uint64_t val, unsigned count);
-uint64_t gen_mask(uint64_t width);
+uint64_t reflect(uint64_t val, uint8_t count);
+uint64_t gen_mask(uint8_t width);
 #endif /* SIMPLECRC_UTILITY_H_ */
